+++ domain/src/main/kotlin/com/yapp/bol/file/FileInfo.kt
@@ -9,12 +9,3 @@ interface FileInfo {
 
     fun getUrl(): String
 }
-
-data class MockFileInfo(
-    override val id: FileId,
-    override val uuid: String,
-) : FileInfo {
-    constructor() : this(FileId(0), "uuid")
-
-    override fun getUrl(): String = uuid
-}
