+++ domain/src/test/kotlin/com/yapp/bol/file/MockFileInfo.kt
@@ -1,0 +1,10 @@
+package com.yapp.bol.file
+
+data class MockFileInfo(
+    override val id: FileId,
+    override val uuid: String,
+) : FileInfo {
+    constructor() : this(FileId(0), "uuid")
+
+    override fun getUrl(): String = uuid
+}
